+++ src/main/java/net/soulsweaponry/util/ItemUpgradeRecipeJsonBuilder.java
@@ -26,7 +26,9 @@ import net.soulsweaponry.recipe.ItemUpgradeRecipe;
  *   "base": { ... },
  *   "addition": { ... },
  *   "primaryBonus": <float>, // Often used as melee damage, ranged damage, armor bonus, etc.
- *   "secondaryBonus": <float> // Often used as attack speed, draw haste, armor toughness, etc.
+ *   "secondaryBonus": <float> // Often used as attack speed, draw haste, armor toughness, mining efficiency, etc.
+ *   "fallback": <boolean> // True if other upgrade recipes should be prioritized over this one
+ *      // (i.e. upgrade for diamond sword specifically should be chosen over generic tag recipes)
  * }
  */
 public class ItemUpgradeRecipeJsonBuilder {
@@ -36,35 +38,42 @@ public class ItemUpgradeRecipeJsonBuilder {
     private final RecipeCategory category;
     private final float primaryBonus;
     private final float secondaryBonus;
+    private final boolean fallback;
 
     private final Map<String, AdvancementCriterion<?>> criteria = new LinkedHashMap<>();
 
-    private ItemUpgradeRecipeJsonBuilder(Ingredient template, Ingredient base, Ingredient addition, RecipeCategory category, float primaryBonus, float secondaryBonus) {
+    private ItemUpgradeRecipeJsonBuilder(Ingredient template, Ingredient base, Ingredient addition, RecipeCategory category, float primaryBonus, float secondaryBonus, boolean fallback) {
         this.template = template;
         this.base = base;
         this.addition = addition;
         this.category = category;
         this.primaryBonus = primaryBonus;
         this.secondaryBonus = secondaryBonus;
+        this.fallback = fallback;
     }
 
-    public static ItemUpgradeRecipeJsonBuilder create(Ingredient template, Ingredient base, Ingredient addition, RecipeCategory category, float primaryBonus, float secondaryBonus) {
-        return new ItemUpgradeRecipeJsonBuilder(template, base, addition, category, primaryBonus, secondaryBonus);
+    public static ItemUpgradeRecipeJsonBuilder create(Ingredient template, Ingredient base, Ingredient addition, RecipeCategory category, float primaryBonus, float secondaryBonus, boolean fallback) {
+        return new ItemUpgradeRecipeJsonBuilder(template, base, addition, category, primaryBonus, secondaryBonus, fallback);
     }
 
     /** Convenience: items everywhere */
     public static ItemUpgradeRecipeJsonBuilder create(ItemConvertible template, ItemConvertible base, ItemConvertible addition, RecipeCategory category, float primaryBonus, float secondaryBonus) {
-        return create(Ingredient.ofItems(template), Ingredient.ofItems(base), Ingredient.ofItems(addition), category, primaryBonus, secondaryBonus);
+        return create(Ingredient.ofItems(template), Ingredient.ofItems(base), Ingredient.ofItems(addition), category, primaryBonus, secondaryBonus, false);
+    }
+
+    /** Convenience: items everywhere */
+    public static ItemUpgradeRecipeJsonBuilder create(ItemConvertible template, ItemConvertible base, ItemConvertible addition, RecipeCategory category, float primaryBonus, float secondaryBonus, boolean fallback) {
+        return create(Ingredient.ofItems(template), Ingredient.ofItems(base), Ingredient.ofItems(addition), category, primaryBonus, secondaryBonus, fallback);
     }
 
     /** Convenience: tag for base */
-    public static ItemUpgradeRecipeJsonBuilder create(ItemConvertible template, TagKey<Item> baseTag, ItemConvertible addition, RecipeCategory category, float primaryBonus, float secondaryBonus) {
-        return create(Ingredient.ofItems(template), Ingredient.fromTag(baseTag), Ingredient.ofItems(addition), category, primaryBonus, secondaryBonus);
+    public static ItemUpgradeRecipeJsonBuilder create(ItemConvertible template, TagKey<Item> baseTag, ItemConvertible addition, RecipeCategory category, float primaryBonus, float secondaryBonus, boolean fallback) {
+        return create(Ingredient.ofItems(template), Ingredient.fromTag(baseTag), Ingredient.ofItems(addition), category, primaryBonus, secondaryBonus, fallback);
     }
 
     /** Convenience: tags for base/addition */
-    public static ItemUpgradeRecipeJsonBuilder create(ItemConvertible template, TagKey<Item> baseTag, TagKey<Item> additionTag, RecipeCategory category, float primaryBonus, float secondaryBonus) {
-        return create(Ingredient.ofItems(template), Ingredient.fromTag(baseTag), Ingredient.fromTag(additionTag), category, primaryBonus, secondaryBonus);
+    public static ItemUpgradeRecipeJsonBuilder create(ItemConvertible template, TagKey<Item> baseTag, TagKey<Item> additionTag, RecipeCategory category, float primaryBonus, float secondaryBonus, boolean fallback) {
+        return create(Ingredient.ofItems(template), Ingredient.fromTag(baseTag), Ingredient.fromTag(additionTag), category, primaryBonus, secondaryBonus, fallback);
     }
 
     public ItemUpgradeRecipeJsonBuilder criterion(String name, AdvancementCriterion<?> criterion) {
@@ -86,7 +95,7 @@ public class ItemUpgradeRecipeJsonBuilder {
 
         this.criteria.forEach(adv::criterion);
 
-        ItemUpgradeRecipe recipe = new ItemUpgradeRecipe(this.template, this.base, this.addition, this.primaryBonus, this.secondaryBonus);
+        ItemUpgradeRecipe recipe = new ItemUpgradeRecipe(this.template, this.base, this.addition, this.primaryBonus, this.secondaryBonus, this.fallback);
         exporter.accept(id, recipe, adv.build(id.withPrefixedPath("recipes/" + this.category.getName() + "/")));
     }